+++ roborazzi-accessibility-check/src/main/java/com/github/takahirom/roborazzi/ATFAccessibilityChecker.kt
@@ -4,23 +4,65 @@ import android.annotation.SuppressLint
 import android.graphics.Bitmap
 import android.os.Build
 import android.view.View
+import androidx.annotation.RequiresApi
 import androidx.compose.ui.platform.ViewRootForTest
 import androidx.test.espresso.UiController
 import androidx.test.espresso.ViewAction
 import com.github.takahirom.roborazzi.RoborazziRule.CaptureRoot
+import com.google.android.apps.common.testing.accessibility.framework.AccessibilityCheckPreset
 import com.google.android.apps.common.testing.accessibility.framework.AccessibilityCheckResult.AccessibilityCheckResultType
 import com.google.android.apps.common.testing.accessibility.framework.AccessibilityHierarchyCheck
 import com.google.android.apps.common.testing.accessibility.framework.AccessibilityViewCheckResult
+import com.google.android.apps.common.testing.accessibility.framework.Parameters
+import com.google.android.apps.common.testing.accessibility.framework.ViewChecker
 import com.google.android.apps.common.testing.accessibility.framework.integrations.espresso.AccessibilityViewCheckException
+import com.google.android.apps.common.testing.accessibility.framework.utils.contrast.BitmapImage
+import com.google.common.collect.ImmutableSet
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers
 import org.robolectric.shadows.ShadowBuild
 
 @ExperimentalRoborazziApi
 data class ATFAccessibilityChecker(
-  val checks: Set<AccessibilityHierarchyCheck>,
-  val suppressions: Matcher<in AccessibilityViewCheckResult>,
+  val checks: Set<AccessibilityHierarchyCheck> = AccessibilityCheckPreset.getAccessibilityHierarchyChecksForPreset(
+    AccessibilityCheckPreset.LATEST
+  ),
+  val suppressions: Matcher<in AccessibilityViewCheckResult> = Matchers.not(Matchers.anything()),
 ) {
+  constructor(
+    preset: AccessibilityCheckPreset,
+    suppressions: Matcher<in AccessibilityViewCheckResult> = Matchers.not(Matchers.anything()),
+  ) : this(
+    checks = AccessibilityCheckPreset.getAccessibilityHierarchyChecksForPreset(preset),
+    suppressions = suppressions,
+  )
+
+
+  @RequiresApi(34)
+  internal fun runAllChecks(
+    view: View,
+    screenshotBitmap: Bitmap?,
+    checks: Set<AccessibilityHierarchyCheck>,
+    suppressions: Matcher<in AccessibilityViewCheckResult>,
+  ): List<AccessibilityViewCheckResult> {
+    val parameters = Parameters().apply {
+      if (screenshotBitmap != null) {
+        putScreenCapture(BitmapImage(screenshotBitmap))
+      }
+      setSaveViewImages(true)
+    }
+
+    val viewChecker = ViewChecker().apply {
+      setObtainCharacterLocations(true)
+    }
+
+    val results = viewChecker.runChecksOnView(ImmutableSet.copyOf(checks), view, parameters)
+
+    return results.filter {
+      !suppressions.matches(it)
+    }
+  }
+
   @SuppressLint("VisibleForTests")
   fun runAccessibilityChecks(
     captureRoot: CaptureRoot,
@@ -31,27 +73,24 @@ data class ATFAccessibilityChecker(
       roborazziErrorLog("Skipping accessibilityChecks on API " + Build.VERSION.SDK_INT + "(< ${Build.VERSION_CODES.UPSIDE_DOWN_CAKE})")
       return
     }
-    // TODO remove this once ATF doesn't bail out
-    // https://github.com/google/Accessibility-Test-Framework-for-Android/blob/c65cab02b2a845c29c3da100d6adefd345a144e3/src/main/java/com/google/android/apps/common/testing/accessibility/framework/uielement/AccessibilityHierarchyAndroid.java#L667
-    ShadowBuild.setFingerprint("roborazzi")
+
+    if (Build.FINGERPRINT == "robolectric") {
+      // TODO remove this once ATF doesn't bail out
+      // https://github.com/google/Accessibility-Test-Framework-for-Android/blob/c65cab02b2a845c29c3da100d6adefd345a144e3/src/main/java/com/google/android/apps/common/testing/accessibility/framework/uielement/AccessibilityHierarchyAndroid.java#L667
+      ShadowBuild.setFingerprint("roborazzi")
+    }
 
     if (captureRoot is CaptureRoot.Compose) {
-      val view =
-        (captureRoot.semanticsNodeInteraction.fetchSemanticsNode().root as ViewRootForTest).view.rootView
+      val view = (captureRoot.semanticsNodeInteraction.fetchSemanticsNode().root as ViewRootForTest).view.rootView
 
       // Will throw based on configuration
 
-      val screenshot: Bitmap? =
-        RoboComponent.Compose(
-          node = captureRoot.semanticsNodeInteraction.fetchSemanticsNode(),
-          roborazziOptions = roborazziOptions
-        ).image
+      val screenshot: Bitmap? = RoboComponent.Compose(
+        node = captureRoot.semanticsNodeInteraction.fetchSemanticsNode(), roborazziOptions = roborazziOptions
+      ).image
 
       val results = runAllChecks(
-        view = view,
-        screenshotBitmap = screenshot,
-        checks = checks,
-        suppressions = suppressions
+        view = view, screenshotBitmap = screenshot, checks = checks, suppressions = suppressions
       )
 
       reportResults(results, failureLevel)
@@ -85,8 +124,7 @@ data class ATFAccessibilityChecker(
   }
 
   private fun reportResults(
-    results: List<AccessibilityViewCheckResult>,
-    failureLevel: CheckLevel
+    results: List<AccessibilityViewCheckResult>, failureLevel: CheckLevel
   ) {
     // Report on any warnings in the log output if not failing
     results.forEach { check ->
@@ -118,8 +156,7 @@ enum class CheckLevel(private vararg val failedTypes: AccessibilityCheckResultTy
   Error(AccessibilityCheckResultType.ERROR),
 
   Warning(
-    AccessibilityCheckResultType.ERROR,
-    AccessibilityCheckResultType.WARNING
+    AccessibilityCheckResultType.ERROR, AccessibilityCheckResultType.WARNING
   ),
 
   LogOnly;
@@ -129,12 +166,11 @@ enum class CheckLevel(private vararg val failedTypes: AccessibilityCheckResultTy
 
 @ExperimentalRoborazziApi
 data class AccessibilityChecksValidate(
-  val checker: ATFAccessibilityChecker = ATFAccessibilityChecker.atf(),
+  val checker: ATFAccessibilityChecker = ATFAccessibilityChecker(),
   val failureLevel: CheckLevel = CheckLevel.Error,
 ) : RoborazziRule.AccessibilityChecks {
   override fun runAccessibilityChecks(
-    captureRoot: CaptureRoot,
-    roborazziOptions: RoborazziOptions
+    captureRoot: CaptureRoot, roborazziOptions: RoborazziOptions
   ) {
     checker.runAccessibilityChecks(
       captureRoot = captureRoot,
